fix(sync,nodes): preserve ordering and close v2 sync gaps
- backend move_node endpoint computes fractional sequence from position index

- v2 sync handles move_node, set_classes, set_tags; set_collapsed is local-only

- optimistic create uses sequence to place afterBlockId

- mutationMap fallback uses runtime orderIndex

- type consistency for fractional sequences

+++ app/features/nodes/port.py
@@ -551,7 +551,7 @@ class NodeClassRepository(ABC):
         pass
 
     @abstractmethod
-    async def get_node_sequence(self, node_id: int) -> int | None:
+    async def get_node_sequence(self, node_id: int) -> float | None:
         """Get the sequence of a node."""
         pass
